requirements.txt
Flask==1.1.2
umap-learn==0.5.1
hdbscan==0.8.27 